requirements.txt
# certifi==2026.1.4
# charset-normalizer==3.4.4
# grpcio==1.76.0
# grpcio-tools==1.76.0
# idna==3.11
paho-mqtt==2.1.0
prometheus_client==0.24.1
protobuf==6.32.0
requests==2.32.5
# setuptools==80.9.0
# typing_extensions==4.15.0
# urllib3==2.6.3
# voluptuous==0.16.0
